requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
boto3==1.34.84
python-multipart==0.0.9
python-dotenv==1.0.1
